Refactor Oryx to use Task instead of Async
This is the main feedback I got from the F# community in Antwerp. Task
is 10 times faster than Async and we also don't have to convert back and
forth between Task and Async all the time.The challenge here is that
Task is eager and starts right away,  while Async is lazy and needs to
be started explicitly.

Add proper cancellation support so we can cancel ongoing requests.