Handle empty assembly location (#90)
See https://github.com/cognitedata/cognite-sdk-dotnet/pull/248/files, the issue still occurs since it is also present in Oryx. This is the same fix, which is confirmed to work even if there is no physical assembly. It may also be a bit more flexible, since InformationalVersion has no restrictions on structure.